browser_replica_app/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
langchain==0.1.0
langchain-google-genai==0.0.6
google-genai==1.29.0
langgraph==0.0.20
python-multipart==0.0.6
aiohttp==3.9.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
beautifulsoup4==4.12.2
brotli==1.2.0
feedparser==6.0.10
newspaper4k==0.9.5
gnews==0.4.2
lxml[html_clean]==5.2.2
nltk==3.6.7
pillow==10.4.0
python-dateutil==2.8.2
pyyaml==6.0.1
requests==2.31.0
tldextract==5.1.2
typing-extensions==4.12.2
w3lib==2.3.1
yt-dlp==2026.3.17
